+++ CHANGELOG.md
@@ -182,4 +182,5 @@ Providing multiple input files will only work in `--modify`/`-m` mode.
 
 - Support filenames with spaces (#416, #428)
 - Do not flatten braces which contain commands (#407, #423)
+- Do not encode special characters in verbatim fields like url (#415)
 - Upgrade dependencies